grpc: only check for stale db on empty data returned
This should be safe since if there is any data present it is because it
was from a Push or ingested already. Only when we have "" must we check
if its "not-ingested-yet" or "deleted/not-exists" and so must wait for
the DB.

There is a slight alteration in the metrics gathered. Previously
cacheDuration was not being observed unless the db was ready, which is
now different. This seems fine since it should only be for a little
while at startup, and the current behavior seems like the correct
behavior and we should be measuring it.