deploy: clean up staging-env
Theres quite the refactor going on here but not much actually changes
and when it does I think it's for better readability. Added a vim-mode
line in a comment so we vimmers can have nice syntax highlighting.

I made it so there is not a group of vars that get broken up in the
middle by template syntax, so a block/group looks like a block/group.

Saving the env into a variable makes more sense than calling out to the
`env` function for each use since it avoids "parsing" a function call.

Another important change is to use pipeline more, since I think they
read better as one line for the `with` instead of one to setup a var and
one for the secret.