main: ingest facility data asap
No longer a need to wait for a push or an explicit rpc call. Waiting to
ingest is plainly just the wrong thing to do now. Less active (re
provisions) facilities will not get a push and so the BMCs will fail to
DHCP while "DB is not ready", and in hashistack cacher seems to be
restarted somewhat more frequently.