ingest: refactor fetchFacility and copyin to run concurrently
This is where we take advantage of the refactor to copyin in last commit
that changes the copyin from one huge []map[string]interface{} to one
per page of facility data.

This has been possible for a long time now, since I added the check for
pre-existing row via id to handle concurrent Push/Ingest. A Push will
continue to not be overriden by fetched data and in the unlikely case of
fetched data being newer than a push, we will soon get the newer push
anyway. This makes it so that just like a Push, if we have data to serve
to a client then there's no need to wait for *all* the data to be
ingested.