Avoid deadlock if fetch or copyin encounters an error
If there's an actual error in either fetch or copyin the goroutine will
block waiting for the reader since it was an unbuffered channel. But the
caller is waiting on `wg.Wait()` before it tries to read any errors,
which will never happen while the errored goroutine is blocked on
sending. Deadlock. But we don't exit or fail or retry.