Replace DB stuff with hardware
We aren't really using the db for much as is, at most just for the
indexing and query from indexed fields. But its a pain to have the
conceptual overhead of an external DB when we need to overwrite all the
data at start up anyway.

Having everything in memory is faster than going out to a DB even if on
the same node and also makes HA a little closer to possible.