make sure the close happens within the watchLock
we've seen issues where a client can lose connection and attempt a
reconnect before the stream has been detected as done. So we hit
a condition where we have broken out of the select due to a disconnect
but the new connection has already attempted to evict the old channel.